Generate new shared-context dist
