--- external_modules/shared-context/walcron-zelda-shared-context.js
@@ -1,2 +1,0 @@
-System.register(["rxjs"],(function(e,o){var r={};return{setters:[function(e){r.BehaviorSubject=e.BehaviorSubject}],execute:function(){e((()=>{var e={722:(e,o,r)=>{const t=r(905).R;o.s=function(e){if(e||(e=1),!r.y.meta||!r.y.meta.url)throw console.error("__system_context__",r.y),Error("systemjs-webpack-interop was provided an unknown SystemJS context. Expected context.meta.url, but none was provided");r.p=t(r.y.meta.url,e)}},905:(e,o,r)=>{o.R=function(e,o){var r=document.createElement("a");r.href=e;for(var t="/"===r.pathname[0]?r.pathname:"/"+r.pathname,n=0,s=t.length;n!==o&&s>=0;)"/"===t[--s]&&n++;if(n!==o)throw Error("systemjs-webpack-interop: rootDirectoryLevel ("+o+") is greater than the number of directories ("+n+") in the URL path "+e);var i=t.slice(0,s+1);return r.protocol+"//"+r.host+i};Number.isInteger},792:e=>{"use strict";e.exports=r}},t={};function n(o){var r=t[o];if(void 0!==r)return r.exports;var s=t[o]={exports:{}};return e[o](s,s.exports,n),s.exports}n.y=o,n.d=(e,o)=>{for(var r in o)n.o(o,r)&&!n.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:o[r]})},n.o=(e,o)=>Object.prototype.hasOwnProperty.call(e,o),n.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.p="";var s={};return(0,n(722).s)(1),(()=>{"use strict";n.r(s),n.d(s,{INTERVAL_CHECK_IN_MILISECONDS:()=>r,auth$:()=>t,login:()=>a,logout:()=>u,name:()=>c});var e=n(792),o="abc1234",r=2500,t=new e.BehaviorSubject({sessionToken:localStorage.getItem("sessionToken"),error:void 0,pending:!1}),i=function(e,n){return new Promise((function(s){t.next({sessionToken:null,error:void 0,pending:!0}),setTimeout((function(){"walcoorperation@gmail.com"===e&&"samplePassword"===n?(localStorage.setItem("sessionToken",o),s({sessionToken:o,error:void 0,pending:!1})):s({sessionToken:null,error:"Invalid user or password",pending:!1})}),r)}))};function a(e,o){t.value.pending||i(e,o).then((function(e){t.next(e)}))}function u(){localStorage.removeItem("sessionToken"),t.next({sessionToken:null,error:void 0,pending:!1})}var c="shared-context"})(),s})())}}}));
-//# sourceMappingURL=walcron-zelda-shared-context.js.map
